Adding struct_data.py to the repository. This file contains functions that are used to manipulate dataframes in the battery data analysis project. The functions include adding a column identifier, converting a column to a string, concatenating dataframes, classifying dataframes, and calculating the total seconds of resting the battery experiences per cycle. The functions are used to clean and organize the data for analysis.
Additional changes include updating the README.md file to include
a description of the struct_data.py file and the functions it contains.
This function still requires data condensation and cleaning to be
added to the aging model. Once fully complete, these data will be able
to be used to train the model and predict battery life.